requirements.txt
altgraph==0.17
anyio==2.2.0
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
autopep8==1.5.5
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.7
dnspython==2.1.0
email-validator==1.1.2
entrypoints==0.3
Flask==1.1.2
flask-mongoengine==1.0.0
Flask-Session==0.3.2
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.0.0
gunicorn==20.1.0
idna==2.8
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.0
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.3.0
keyboard==0.13.5
MarkupSafe==1.1.1
mistune==0.8.4
mongoengine==0.23.0
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
oauthlib==3.0.1
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==8.1.0
playsound==1.2.2
prometheus-client==0.9.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pygame==2.0.1
pyglet==1.5.15
Pygments==2.8.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pymongo==3.11.3
pynput==1.7.3
PyOpenGL @ file:///C:/Users/Natnael%20Abay%20Akalu/AppData/Local/Programs/Python/Python38/PyOpenGL-3.1.5-cp38-cp38-win_amd64.whl
PyOpenGL-accelerate @ file:///C:/Users/Natnael%20Abay%20Akalu/AppData/Local/Programs/Python/Python38/PyOpenGL_accelerate-3.1.5-cp38-cp38-win_amd64.whl
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
PyWavefront==1.3.3
pywin32==300
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==22.0.3
requests==2.21.0
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.4.5
sqlparse==0.4.1
terminado==0.9.3
testpath==0.4.4
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.24.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
